refactor game setup logic
