+++ src/app/constants/game.js
@@ -1,0 +1,8 @@
+export const DIRECTIONS = [
+  { name: "LEFT", h: -1, v: 0 },
+  { name: "RIGHT", h: +1, v: 0 },
+  { name: "UP", h: 0, v: +1 },
+  { name: "DOWN", h: 0, v: -1 },
+];
+
+export const EMPTY = "-----";
