+++ src/app/helpers/random.js
@@ -1,3 +1,5 @@
+// integers between 0 (inclusive) and num (exclusive)
+// => [ 0 , num )
 export const randomNumber = num => Math.floor(Math.random() * Math.floor(num));
 
 export const randomShuffle = array =>
@@ -5,3 +7,9 @@ export const randomShuffle = array =>
     .map(a => [Math.random(), a])
     .sort((a, b) => a[0] - b[0])
     .map(a => a[1]);
+
+export const randomString = size =>
+  // size-character random string
+  Math.random()
+    .toString(36)
+    .slice(2, 2 + size);
